requirements.txt
manim==0.17.3
anthropic==0.7.7
python-dotenv==1.0.0 